added split tuple function with unit test
